[Feature] Updated models
